agent-worker/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
browser-use